move all item field and section modules under op_items.fields_sections
