+++ pyonepassword/api/exceptions.py
@@ -1,6 +1,6 @@
-from ..op_items._new_fields import OPNewTOTPUriException
 from ..op_items._op_item_type_registry import OPUnknownItemTypeException
-from ..op_items.item_section import (
+from ..op_items.fields_sections._new_fields import OPNewTOTPUriException
+from ..op_items.fields_sections.item_section import (
     OPItemFieldCollisionException,
     OPSectionCollisionException
 )
