+++ pyonepassword/api/object_types.py
@@ -1,7 +1,8 @@
 from ..account import OPAccount
 from ..op_items._item_list import OPItemList
 from ..op_items._new_field_registry import OPNewItemField
-from ..op_items._new_fields import (
+from ..op_items._new_item import OPNewSection
+from ..op_items.fields_sections._new_fields import (
     OPNewConcealedField,
     OPNewPasswordField,
     OPNewStringField,
@@ -9,10 +10,13 @@ from ..op_items._new_fields import (
     OPNewTOTPUri,
     OPNewUsernameField
 )
-from ..op_items._new_item import OPNewSection
-from ..op_items.item_field import OPConcealedField, OPStringField, OPTOTPField
-from ..op_items.item_field_base import OPItemField
-from ..op_items.item_section import OPSection
+from ..op_items.fields_sections.item_field import (
+    OPConcealedField,
+    OPStringField,
+    OPTOTPField
+)
+from ..op_items.fields_sections.item_field_base import OPItemField
+from ..op_items.fields_sections.item_section import OPSection
 from ..op_items.item_types._item_base import OPAbstractItem
 from ..op_items.item_types.api_credential import OPAPICredentialItem
 from ..op_items.item_types.credit_card import OPCreditCardItem
