+++ pyonepassword/op_items/_new_item.py
@@ -5,8 +5,11 @@ from typing import Dict, List, Optional
 
 from ..py_op_exceptions import OPInvalidItemException
 from ._new_field_registry import OPNewItemField, OPNewItemFieldFactory
-from .item_field_base import OPItemField
-from .item_section import OPSection, OPSectionCollisionException
+from .fields_sections.item_field_base import OPItemField
+from .fields_sections.item_section import (
+    OPSection,
+    OPSectionCollisionException
+)
 from .template_directory import OPTemplateDirectory
 from .uuid import OPUniqueIdentifierBase32, is_uuid
